Create main file that handles simulation
Creating main file that handles calculating the gravitational forces
between two bodies. The program prompts the user for a frametime and
number of frames to execute. The program currently defaults to have two
bodies with a mass, initial position, and initial velocity. The gravity
is calculated using Newton's Universal Law of Gravitation.

- The Gravitational Constant is set to 1 to avoid doing math on
  extremely large numbers