Add third body to the simulation
Adding a third body that interacts with the previous two. Eventually
will refine code to allow to more than three bodies. The next goal is to
have a interactive UI to easily create a body with a mass, and placement
on a grid will account for its starting position, and a velocity vector
to easily set initial velocity.